+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   [#3120](https://github.com/oshi/oshi/pull/3120),
   [#3121](https://github.com/oshi/oshi/pull/3121): Port Linux JNA classes to FFM - [@dbwiddis](https://github.com/dbwiddis).
 * [#3122](https://github.com/oshi/oshi/pull/3122): Refactor USB device classes to reduce duplication - [@dbwiddis](https://github.com/dbwiddis).
+* [#3124](https://github.com/oshi/oshi/pull/3124): Fix Windows battery remaining capacity percent not reflecting IOCTL values - [@dbwiddis](https://github.com/dbwiddis).
 
 # 6.11.0 (2026-04-04)